refactor: use a common Text component for help & error
